Remove legacy test-only ResearchViews render methods; fold onto ProjectMember
Following the SourceLine currency adoption (#2757), every real product source
and IL renderer already routes through the unified ResearchViews.ProjectMember
entry point. The remaining RenderCostOverlay / RenderCostOverlayWithHeaderFacts
/ RenderSemanticsOverlay / RenderAnnotatedSource / RenderMixed /
ProjectAnnotatedIl methods had no product callers — only tests — so they are
removed along with the now-dead AddFactsToAnnotatedIl / TryReadOffset helpers.

Tests migrate onto ProjectMember:
- ResearchFactRegistryTests gains test-local overlay/annotated helpers. Two
  tests that asserted cost/semantics facts are absent from the annotated view
  are deleted: they exercised the removed RenderAnnotatedSource path, which
  built its fact context WITHOUT the ResearchAssemblyContext and so never saw
  cross-method facts. The real (CLI) annotated view resolves the assembly and
  does co-render those facts, so the assertions never matched product behavior.
- AnnotatedCSharpRendererTests and AnnotatedIlFactTests both drove the same
  mixed view as MixedSourceRendererTests; they are folded into it, dropping
  three duplicate assertions.

RTS (CompileBackSourceComposer) composes a whole compilable unit via
CSharpPrinter.PrintRaised and stays intentionally off the line currency.

Copilot-Session: d6c5d668-eb01-4f5b-bc22-09e6eade8835
